requirements.txt
allure-pytest==2.13.5
allure-python-commons==2.13.5
annotated-types==0.7.0
attrs==24.3.0
certifi==2024.12.14
charset-normalizer==3.4.1
exceptiongroup==1.2.2
execnet==2.1.1
Faker==33.3.0
idna==3.10
iniconfig==2.0.0
packaging==24.2
pluggy==1.5.0
pydantic==2.10.5
pydantic_core==2.27.2
pytest==8.3.4
pytest-xdist==3.6.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.3
six==1.17.0
tomli==2.2.1
typing_extensions==4.12.2
urllib3==2.3.0
